Code refactor with useCreateMCAccount hook.
